requirements.txt
ansible==4.9.0
chardet==4.0.0
click==8.1.3
coloredlogs==15.0.1
deepdiff==5.8.1
dictdiffer==0.9.0
Django>=4.0.1
django-bootstrap5==22.1
genie==22.5
hmnet==0.1.35
jsonschema==4.16.0
MarkupSafe==1.1.1
nested-lookup==0.2.25
netmiko==3.4.0
pyats==22.5
pysnow==0.7.17
python-dotenv==0.21.0
PyYAML==6.0
pyzscaler==1.3.0
tabulate~=0.8
textfsm==1.1.2
viptela==0.3.9
